fix: renaming entities -> entity
