+++ src/main/java/controly/dto/PostagemDTO.java
@@ -1,9 +1,8 @@
 package controly.dto;
 
-import controly.entities.UserEntity;
-import controly.entities.PostEntity;
-import controly.entities.TopicEntity;
-import lombok.Data;
+import controly.entity.UserEntity;
+import controly.entity.PostEntity;
+import controly.entity.TopicEntity;
 
 import javax.persistence.Id;
