+++ src/main/java/controly/entity/DoubtsAnswerEntity.java
@@ -1,7 +1,6 @@
-package controly.entities;
+package controly.entity;
 
 import lombok.AllArgsConstructor;
-import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
