+++ src/main/java/controly/entity/PostEntity.java
@@ -1,9 +1,8 @@
-package controly.entities;
+package controly.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
-import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.lang.Nullable;
