+++ src/main/java/controly/entity/TopicHasFollowersEntity.java
@@ -1,7 +1,7 @@
-package controly.entities;
+package controly.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.Data;
+
 import javax.persistence.*;
 
 import static javax.persistence.GenerationType.IDENTITY;
