+++ src/main/java/controly/entity/UserEntity.java
@@ -1,8 +1,6 @@
-package controly.entities;
+package controly.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.Data;
-import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
