+++ src/main/java/controly/repository/TopicRepository.java
@@ -1,6 +1,6 @@
 package controly.repository;
 
-import controly.entities.TopicEntity;
+import controly.entity.TopicEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TopicRepository extends JpaRepository<TopicEntity,Long> {
