+++ src/main/java/controly/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package controly.repository;
 
-import controly.entities.UserEntity;
+import controly.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
