+++ src/main/java/controly/strategy/Postagem.java
@@ -1,9 +1,9 @@
 package controly.strategy;
 
-import controly.entities.CommentEntity;
-import controly.entities.PostEntity;
-import controly.entities.TopicEntity;
-import controly.entities.UserEntity;
+import controly.entity.CommentEntity;
+import controly.entity.PostEntity;
+import controly.entity.TopicEntity;
+import controly.entity.UserEntity;
 
 public abstract class Postagem {
